+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   utils: Added currency conversions from any denominations to smallest (Atto)
+
 ### Changed
 
 -   client: `getAccount` now accepts a second optional parameter `fromPool: bool` which will get data about the account from the TxPool's `ethState` instead of the default main `ethState`